+++ CHANGELOG.md
@@ -5,3 +5,11 @@
 ## 0.0.2
 
 * Changed description and removed license from read me
+
+## 0.0.7
+
+* Fixed the length issue for method builder 
+
+## 0.0.8
+
+* Removed deps on secure storage